% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\title{Create embeddings with the arborist package}
\usage{
embed(data, model, type = "binary")
}
\arguments{
\item{data}{`data.frame` or `matrix` - data which will be
embedded by a model}

\item{model}{a tree-based model that will be used to create embeddings.
If this object is an explainer, then a model will be extracted.}

\item{type}{a character with the name of the embedding procedure.
one out of `binary`/`shap`/`bayesian`.
If `binary` then the resulting embedding will have as many dimensions as
we have trees in the forest Each tree will contribute to a single dimension.
The embedding procedure is described in the ,,All you need are trees'' ebook.
If `shap` then Shapley values are used as embeddings. The size of
the embedding corresponds to the number of variables in the 'model'.}
}
\value{
a matrix with embeddings created for observations from the 'data'
}
\description{
The `embed()` function create an embedding for data set
with the use of a random forest model.
Two different approaches to embeddings are implemented.
One is based on SHAP values while the second is a binary encoding
of data flow through the forest.
}
\examples{
library("DALEX")
model <- train(titanic_imputed, "survived", engine = "randomForest", ntree = 25)
embeddings_binary <- embed(titanic_imputed, model, type = "binary")
str(embeddings_binary)

model_fr <- train(small_titanic, "survived", engine = "randomForest", type = "regression")
embeddings_shap   <- embed(titanic_imputed, model, type = "shap")
str(embeddings_shap)

}
