% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Explain forester model}
\usage{
explain(model, ...)
}
\arguments{
\item{model}{a model created with the `train()` function}

\item{...}{other parameters}
}
\value{

}
\description{
The `explain()` is a wrapper for `DALEX` methods for model explanations.
If possible it will use methods for tree based models.
}
\examples{
library(DALEX)
library(magrittr)
titanic_imputed \%>\%
    train("survived") \%>\%
    forester::explain()

}
