% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract knowledge from forester models}
\usage{
extract(model, ...)
}
\arguments{
\item{model}{a model created with the `train()` function}

\item{...}{other parameters}
}
\value{

}
\description{
The `extract()` function extracts some useful statistics, like important features
i.e. these used by the model, ranking of feature importances.
}
\examples{
library(DALEX)
library(magrittr)
titanic_imputed \%>\%
    train("survived") \%>\%
    forester::extract()

}
