% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{get_embedding_forest_binary}
\alias{get_embedding_forest_binary}
\title{Calculate binary embeddings for a random forest model}
\usage{
get_embedding_forest_binary(data, forest, length_of_embedding = 8)
}
\description{
Calculate binary embeddings for a random forest model
}
