% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{get_embedding_forest_shap}
\alias{get_embedding_forest_shap}
\title{Calculate shap embeddings for a random forest model}
\usage{
get_embedding_forest_shap(data, forest, ...)
}
\description{
Calculate shap embeddings for a random forest model
}
