% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{get_embedding_tree_binary}
\alias{get_embedding_tree_binary}
\title{Calculate binary embeddings for a tree from forest model}
\usage{
get_embedding_tree_binary(data, tree, length_of_embedding = 8)
}
\description{
Calculate binary embeddings for a tree from forest model
}
