% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot forester models}
\usage{
plot(model, ...)
}
\arguments{
\item{model}{a model created with the `train()` function}

\item{...}{other parameters}
}
\value{
ggplot2 object
}
\description{
The generic `plot()` function creates and prints numeric summary
of forester models.
See https://cran.r-project.org/web/packages/randomForestExplainer/index.html
and https://github.com/ModelOriented/EIX
for some examples.
}
\examples{
library(DALEX)
library(magrittr)
titanic_imputed \%>\%
    train("survived") \%>\%
    forester::plot()

}
