% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.forester}
\alias{print.forester}
\title{Print forester models}
\usage{
\method{print}{forester}(x, ...)
}
\arguments{
\item{x}{a model created with the `train()` function}

\item{...}{other parameters}
}
\value{
side effect - prints something on the screen
}
\description{
The generic `print()` function creates and prints numeric sumamry
of forester models.
}
\examples{
library(DALEX)
library(magrittr)
titanic_imputed \%>\%
    train("survived")

}
