% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{Print forester models}
\usage{
print(model, ...)
}
\arguments{
\item{model}{a model created with the `train()` function}

\item{...}{other parameters}
}
\value{

}
\description{
The generic `print()` function creates and prints numeric sumamry
of forester models.
}
\examples{
library(DALEX)
library(magrittr)
titanic_imputed \%>\%
    train("survived") \%>\%
    forester::print()

}
