% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train models with forester}
\usage{
train(
  data,
  y,
  data_test = NULL,
  y_test = NULL,
  train_test_ratio = 4,
  type = "guess",
  engine = c("randomForest", "ranger", "xgboost", "lightgbm"),
  loss = "default",
  validation = "default",
  tuning = "default",
  keep = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{`data.frame` or `matrix` - data which will be
used to build models. By default model will be trained
on all columns in the `data`}

\item{y}{target variable. It can be either
(1) a vector of the same number of observations as `data` or
(2) a character name of variable in the `data` that contains
the target variable}

\item{data_test}{data that will be used for evaluation}

\item{y_test}{data that will be used for evaluation}

\item{train_test_ratio}{if data_test is not provided then the orifinal data will be divided in train/test}

\item{type}{a character, one of `classification`/`regression`/`guess`.
sets the type of the task. If `guess` (the default option) then
forester will figure out `type` based on the number of unique values
in the `y` variable}

\item{engine}{a vector of tree-based models that shall be testes.
Possible values are: `ranger`, `xgboost`, `lightgbm`, `catboost`.
All models from this vector will be trained and the best one will be returned.}

\item{loss}{a character with the name of the loss function.
If `default` then the default loss function for a given task will be used}

\item{validation}{validation/model selection a character of
data.frame. If data.frame then this data will be used for model
assessment and selection. If character, then it will be a name of
the hyperparameter.}

\item{tuning}{a character with the name of the tuning procedure.
one out of `default`/`portfolio`/`bayesian`/`random search`.
If `default` then no tuning, just default hyperparameteres are used.
If `portfolio` then static portfolio of hyperparametes are tried.
If `bayesian` then Bayesian optimization is performed.
If `random search` then random search is performed.}

\item{keep}{shall all models be returned (`keep = TRUE`) or only the best one
(`keep = FALSE`, default option)?}

\item{verbose}{logical. If TRUE (default) then diagnostic messages will be printed}
}
\value{
an DALEX explainer (or set of explainers if keep=TRUE)
}
\description{
The `train()` is the core function of this package.
The only obligatory arguments are `data` and `target`.
Setting and changing other arguments will affect model
validation strategy, tested model families and so on.
}
\examples{
library(DALEX)
train(titanic_imputed, "survived")

}
