% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_randomForest}
\alias{train_randomForest}
\title{Train a simple randomForest model}
\usage{
train_randomForest(
  data,
  y,
  type = "guess",
  loss = "default",
  validation = "default",
  tuning = "default",
  label = "Random Forest"
)
}
\arguments{
\item{data}{`data.frame` or `matrix` - data which will be
used to build models. By default model will be trained
on all columns in the `data`}

\item{y}{target variable. It can be either
(1) a vector of the same number of observations as `data` or
(2) a character name of variable in the `data` that contains
the target variable}

\item{type}{a character, one of `classification`/`regression`/`guess`.
sets the type of the task. If `guess` (the default option) then
forester will figure out `type` based on the number of unique values
in the `y` variable}

\item{loss}{see `train` function}

\item{validation}{see `train` function}

\item{tuning}{see `train` function}

\item{label}{label for the model}
}
\value{
an DALEX explainer
}
\description{
Train a simple randomForest model
}
\examples{
library(DALEX)
train_randomForest(titanic_imputed, "survived")

}
