% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_ranger}
\alias{train_ranger}
\title{Train a simple ranger model}
\usage{
train_ranger(
  data,
  y,
  data_test,
  type = "guess",
  loss = "default",
  validation = "default",
  tuning = "default",
  label = "Ranger",
  verbose = FALSE
)
}
\arguments{
\item{data}{see the `train` function}

\item{y}{see the `train` function}

\item{data_test}{see the `train` function}

\item{type}{a character, one of `classification`/`regression`/`guess`.
sets the type of the task. If `guess` (the default option) then
forester will figure out `type` based on the number of unique values
in the `y` variable}

\item{loss}{see `train` function}

\item{validation}{see `train` function}

\item{tuning}{see `train` function}

\item{label}{label for the model}

\item{verbose}{see `train` function}
}
\value{
an DALEX explainer
}
\description{
Train a simple ranger model
}
